Refactor edge JSON api into its own source file
